Rename test_blobstore.py to blobstore_common_tests.py
This way, the unit test runner doesn't find it.
