+++ tests/test_gsblobstore.py
@@ -11,7 +11,7 @@ sys.path.insert(0, pkg_root)  # noqa
 from cloud_blobstore import BlobNotFoundError
 from cloud_blobstore.gs import GSBlobStore
 from tests import infra
-from tests.test_blobstore import BlobStoreTests
+from tests.blobstore_common_tests import BlobStoreTests
 
 
 class TestGSBlobStore(unittest.TestCase, BlobStoreTests):
